+++ examples/bot.php
@@ -28,6 +28,8 @@ use danog\MadelineProto\EventHandler\Filter\FilterCommand;
 use danog\MadelineProto\EventHandler\Filter\FilterRegex;
 use danog\MadelineProto\EventHandler\Filter\FilterText;
 use danog\MadelineProto\EventHandler\Message;
+use danog\MadelineProto\EventHandler\Message\Service\DialogMessagePinned;
+use danog\MadelineProto\EventHandler\Message\Service\DialogPhotoChanged;
 use danog\MadelineProto\EventHandler\SimpleFilter\FromAdmin;
 use danog\MadelineProto\EventHandler\SimpleFilter\Incoming;
 use danog\MadelineProto\Logger;
@@ -178,6 +180,20 @@ class MyEventHandler extends SimpleEventHandler
     public function pingCommand(Incoming&Message $message): void
     {
         $message->reply('hello');
+    }
+
+
+    /**
+     * Called when the dialog photo of a chat or channel changes.
+     */
+    #[Handler]
+    public function logPhotoChanged(Incoming&DialogPhotoChanged $message): void
+    {
+        if ($message->photo) {
+            $message->reply("Nice :D");
+        } else {
+            $message->reply("Aww, why did you delete the group photo? :(");
+        }
     }
 
     /**